moved 3dboard to src folder
